services/agent-gateway/pyproject.toml
[project]
name = "agent-gateway"
version = "0.1.0"
description = "Provider-agnostic agent gateway — OpenAI-compatible API over pluggable runtimes"
readme = "README.md"
authors = [
    { name = "Russell Cox", email = "russ771@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "mlflow>=3.0.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "sse-starlette>=2.0.0",
]

[project.scripts]
agent-gateway = "agent_gateway.main:cli"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
